requirements.txt
git+https://github.com/Rapptz/discord.py.git@5c2431d5435d8a921a83cc05de29968f22a777ef
termcolor==1.1.0
python-dotenv==0.19.2
python-dateutil~=2.8.2
pymongo~=3.12.0
parsedatetime~=2.6
